use more suitable type

+++ drivers/power_typedefs.c
@@ -4,7 +4,7 @@
 struct power_domain_struct
 {
     iopinid_t port;
-    int pin;
+    uint8_t pin; // no port is going to have more than 255 pins...
     void (*pre_enable)(void);
     void (*post_enable)(void);
     void (*pre_disable)(void);
